remove admin-auth
