+++ frontend/app/admin/layout.js
@@ -2,7 +2,7 @@
 // still need to check authentication on every server action!!
 
 import Link from "next/link";
-import { authorizeAdminRoute } from "../_utils/adminSession";
+import { authorizeUserRoute, authorizeAdminRoute } from "../_utils/userSession";
 
 export default async function AdminLayout({ children }) {
   // redirects to /login if not logged in
