+++ frontend/app/profile/layout.js
@@ -8,17 +8,20 @@ export default async function ProfileLayout({ children }) {
   return (
     <>
       <header className="flex flex-wrap items-center justify-between bg-gray-50 p-3">
-        <Link className="font-medium underline" href="/profile">
-          Profile
-        </Link>
+        <div className="flex items-center gap-3">
+          <Link className="font-medium underline" href="/profile">
+            profile
+          </Link>
+          <Link className="font-medium underline" href="/">
+            home
+          </Link>
+        </div>
         <div className="flex items-center gap-3">
           {session.email}
           {session.isAdmin && (
-            <form action="/admin">
-              <button type="submit" className="underline">
-                admin
-              </button>
-            </form>
+            <Link className="font-medium underline" href="/admin">
+              admin
+            </Link>
           )}
           <form action="/login/logout" method="POST">
             <button type="submit" className="underline">
